Rename lightRay.h to lowercase for consistency
It's the only source file in the repository using
camelCase, and it was really bothering me.
