+++ src/accelerators/bvh.c
@@ -15,7 +15,7 @@
 #include "../datatypes/poly.h"
 #include "../datatypes/vector.h"
 #include "../datatypes/bbox.h"
-#include "../datatypes/lightRay.h"
+#include "../datatypes/lightray.h"
 #include "../datatypes/mesh.h"
 #include "../datatypes/sphere.h"
 #include "../datatypes/instance.h"
