+++ src/renderer/pathtrace.h
@@ -10,7 +10,7 @@
 
 #include "../datatypes/vector.h"
 #include "../datatypes/poly.h"
-#include "../datatypes/lightRay.h"
+#include "../datatypes/lightray.h"
 #include "../datatypes/material.h"
 #include "../nodes/bsdfnode.h"
